refactor(p2p): separate runtime and dataset protocol
