+++ core/auki-p2p-dataset/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "auki-p2p-dataset"
+version = "0.1.0"
+edition = "2021"
+rust-version.workspace = true
+publish = false
+description = "Authenticated immutable dataset protocol for auki-p2p"
+readme = "README.md"
+
+[dependencies]
+anyhow = { workspace = true }
+async-trait = { workspace = true }
+auki-p2p = { path = "../auki-p2p" }
+chrono = { workspace = true }
+futures = { workspace = true }
+hex = { workspace = true }
+parking_lot = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+sha2 = { workspace = true }
+thiserror = { workspace = true }
+tokio = { workspace = true, features = ["fs", "io-util", "macros", "rt-multi-thread", "sync", "time"] }
+tokio-util = { workspace = true, features = ["rt"] }
+tracing = { workspace = true }
+uuid = { workspace = true }
+
+[dev-dependencies]
+hickory-resolver = "0.24.4"
+jsonwebtoken = "9.3.0"
+libp2p = { workspace = true, features = ["tokio", "macros", "noise", "tcp", "yamux", "relay", "metrics"] }
+libp2p-stream = { workspace = true }
+prometheus-client = "0.22.3"
+tempfile = { workspace = true }
