+++ core/auki-p2p/src/error.rs
@@ -47,6 +47,8 @@ pub enum Error {
     TransportBuild(String),
     #[error("auki-p2p requires an active Tokio runtime")]
     RuntimeUnavailable,
+    #[error("authenticated application protocol task failed")]
+    ProtocolTask(#[source] tokio::task::JoinError),
     #[error("failed to listen on {address}: {reason}")]
     Listen { address: String, reason: String },
     #[error("failed to open libp2p stream: {0}")]
