+++ core/compute-node-runner-api/examples/relay-file-demo/Cargo.toml
@@ -8,6 +8,7 @@ default-run = "posemesh-relay-file-robot"
 [dependencies]
 anyhow = { workspace = true }
 async-trait = { workspace = true }
+auki-p2p-dataset = { path = "../../../auki-p2p-dataset" }
 chrono = { workspace = true }
 dotenvy = "0.15"
 hex = { workspace = true }
