+++ core/compute-node-runner-api/src/types.rs
@@ -1,11 +1,8 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use std::path::PathBuf;
 use url::Url;
 use uuid::Uuid;
-
-pub const P2P_DATASET_SCHEMA: &str = "auki-p2p-dataset/v0";
 
 /// Lease envelope received from DMS (see SPECS §3.1).
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq)]
@@ -44,30 +41,6 @@ impl LeaseEnvelope {
         lease.p2p_access_token_expires_at = None;
         lease
     }
-}
-
-/// Local ZIP registration accepted by the process-level P2P dataset service.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct P2pDatasetRegistration {
-    pub dataset_id: String,
-    pub domain_id: Uuid,
-    pub name: String,
-    pub path: PathBuf,
-    pub available_until: DateTime<Utc>,
-}
-
-/// Routing and integrity metadata persisted as the small Domain Data reference.
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-pub struct P2pDatasetReference {
-    pub schema: String,
-    pub dataset_id: String,
-    pub domain_id: Uuid,
-    pub name: String,
-    pub peer_id: String,
-    pub multiaddrs: Vec<String>,
-    pub size_bytes: u64,
-    pub sha256: String,
-    pub available_until: DateTime<Utc>,
 }
 
 /// Task specification (see SPECS §3.2).
