+++ core/compute-node/Cargo.toml
@@ -12,6 +12,7 @@ path = "src/lib.rs"
 
 [dependencies]
 auki-p2p = { path = "../auki-p2p" }
+auki-p2p-dataset = { path = "../auki-p2p-dataset" }
 compute-runner-api = { package = "posemesh-compute-node-runner-api", version = "0.1.2", path = "../compute-node-runner-api" }
 anyhow = { workspace = true }
 async-trait = { workspace = true }
@@ -45,12 +46,8 @@ posemesh-domain-http = { workspace = true }
 async-trait = { workspace = true }
 futures = { workspace = true }
 httpmock = { workspace = true }
-hickory-resolver = "0.24.4"
 jsonwebtoken = "9.3.0"
-libp2p = { workspace = true, features = ["tokio", "macros", "noise", "tcp", "yamux", "relay", "metrics"] }
-libp2p-stream = { workspace = true }
 once_cell = { workspace = true }
-prometheus-client = "0.22.3"
 rand = { workspace = true }
 tokio = { workspace = true, features = [
     "rt-multi-thread",
